improve receiver impl

+++ crates/ws-receiver/src/fileio.rs
@@ -1,21 +1,53 @@
 use std::path::Path;
-use log::error;
+use log::{debug, error, info};
 use tokio::fs;
 use tokio::fs::OpenOptions;
 use tokio::io::AsyncWriteExt;
 use ws_common::{AppError, FileEntry, FileMeta, Result};
+use crate::PID_FILE;
+
+/// Prepare output directory
+pub async fn prepare_output_dir(output_dir: &str) -> Result<()> {
+    // check output_dir
+    let out_dir = Path::new(&output_dir);
+    let pid_file = out_dir.join(PID_FILE);
+    if out_dir.exists() {
+        if !out_dir.is_dir() {
+            return Err(AppError::InvalidDir(output_dir.to_string()));
+        }
+        // check if `PID_FILE` file exists
+        info!("pid file: {:?}, exists? {}", pid_file, pid_file.exists());
+        if pid_file.exists() {
+            return Err(AppError::DirInUse(output_dir.to_string()));
+        }
+    } else {
+        // create output_dir
+        fs::create_dir_all(out_dir).await.map_err(AppError::FailedCreateDir)?;
+    }
+
+    // create `PID_FILE` for current receiver
+    fs::write(pid_file, format!("{}", std::process::id())).await
+        .map_err(AppError::FailedWriteFile)?;
+    debug!("[Receiver] Done writing pid file in dir: {:?}", out_dir);
+
+    Ok(())
+}
 
 /// Remote all files and directories in `output_dir`
 pub async fn clear_dir(output_dir: String) -> Result<()> {
+    let pid_file = Path::new(&output_dir).join(PID_FILE);
     let mut entries = fs::read_dir(&output_dir).await.map_err(AppError::FailedReadDir)?;
     while let Some(entry) = entries.next_entry().await.map_err(AppError::DirEntryError)? {
         let path = entry.path();
-        // TODO: skip PID file
         if path.is_dir() {
             if let Err(err) = fs::remove_dir_all(path.as_path()).await {
                 error!("failed to remove dir: {}, error: {}", path.display(), err);
             }
         } else {
+            if path == pid_file {
+                debug!("skip pid file: {:?}", path);
+                continue;
+            }
             if let Err(err) = fs::remove_file(path.as_path()).await {
                 error!("failed to remove file: {}, error: {}", path.display(), err);
             }
